+++ packages/data-provider/src/index.ts
@@ -30,8 +30,6 @@ export * from './types/web';
 export * from './types/graph';
 /* access permissions */
 export * from './accessPermissions';
-/* admin types & capabilities */
-export * from './admin';
 /* query/mutation keys */
 export * from './keys';
 /* api call helpers */
